Refactor: Changed data sent to client from root and changed lookup method of polls.
root: Instead of sending all data to index.html, only id, name and votes of each poll are sent.
votes is removed from database, so countVotes folder finds the total number of votes in each poll now.
on /polls/:poll route: Poll is now looked up based on id not the name.
name, list of options and number of votes are saved as locals for ejs.