Bug: Fixed a bug with updating the page after voting.
Sometimes, after casting a vote, the user receives the same page with the old votes.
This happened because mongo conntecs asynchronously which may lead to the page reload before the database updated.
This is fixed by adding a callback function to update method and have that function call renderPoll.
This way, renderPoll is called after the update.